combine fluid recipe classes and fluid recipe serializer classes
